Move simple utils to _utils
