+++ tests/ethpm/utils/test_contract_utils.py
@@ -3,8 +3,10 @@ import pytest
 from ethpm.exceptions import InsufficientAssetsError, ValidationError
 from ethpm.utils.contract import (
     generate_contract_factory_kwargs,
-    validate_contract_name,
     validate_minimal_contract_factory_data,
+)
+from ethpm.validation import (
+    validate_contract_name,
     validate_w3_instance,
 )
